pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "sift-cli"
version = "0.2.0"
description = "Structured session capture & AI extraction CLI"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.10"
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Software Development :: Documentation",
    "Topic :: Utilities",
]
dependencies = [
    "typer>=0.9.0",
    "rich>=13.0.0",
    "pyyaml>=6.0",
    "jinja2>=3.1.0",
    "python-dotenv>=1.0.0",
    "tomli>=2.0;python_version<'3.11'",
    "tomli-w>=1.0.0",
]

[project.urls]
Homepage = "https://github.com/sirrele/sift"
Repository = "https://github.com/sirrele/sift"
Issues = "https://github.com/sirrele/sift/issues"
Changelog = "https://github.com/sirrele/sift/blob/main/CHANGELOG.md"

[project.optional-dependencies]
anthropic = ["anthropic>=0.18.0"]
gemini = ["google-genai>=1.0.0"]
ollama = ["httpx>=0.27.0"]
pdf = ["pdfplumber>=0.10.0", "pypdf>=3.17.0"]
tui = ["textual>=0.50.0"]
analyze = [
    "tree-sitter>=0.21.0",
    "tree-sitter-languages>=1.10.0;python_version<'3.13'",
]
mcp = ["mcp>=1.0.0"]
keyring = ["keyring>=24.0"]
telemetry = [
    "opentelemetry-api>=1.20.0",
    "opentelemetry-sdk>=1.20.0",
    "opentelemetry-exporter-otlp-proto-grpc>=1.20.0",
]
i18n = ["babel>=2.14"]
all = [
    "anthropic>=0.18.0",
    "google-genai>=1.0.0",
    "httpx>=0.27.0",
    "pdfplumber>=0.10.0",
    "pypdf>=3.17.0",
    "textual>=0.50.0",
    "tree-sitter>=0.21.0",
    "tree-sitter-languages>=1.10.0;python_version<'3.13'",
    "mcp>=1.0.0",
    "keyring>=24.0",
]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
    "ruff>=0.4.0",
    "mypy>=1.8.0",
]

[project.scripts]
sift = "sift.cli:app"
sift-mcp = "sift.mcp.server:main"

[project.entry-points."sift.providers"]
anthropic = "sift.providers.anthropic_provider:AnthropicProvider"
gemini = "sift.providers.gemini_provider:GeminiProvider"
ollama = "sift.providers.ollama_provider:OllamaProvider"

[project.entry-points."sift.analyzers"]
python_ast = "sift.analyzers.python_ast_analyzer:analyze_python_file"

[project.entry-points."sift.output_formatters"]
yaml = "sift.core.build_service:BuildService"

[tool.setuptools.packages.find]
include = ["sift*"]

[tool.setuptools.package-data]
sift = ["locales/*.pot", "locales/*/LC_MESSAGES/*.mo"]

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.mypy]
python_version = "3.10"
warn_return_any = false
warn_unused_configs = true
ignore_missing_imports = true
check_untyped_defs = false
disallow_untyped_defs = false

[tool.ruff]
target-version = "py310"
line-length = 100
